requirements.txt
# Core dependencies
openai>=1.12.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Data processing
pandas>=2.0.0

# Progress bars
tqdm>=4.65.0

# Embeddings and similarity
scikit-learn>=1.3.0

# Chemistry metrics (for ChemCoT datasets)
PyTDC>=1.0.0
rdkit>=2023.3.1

# Logging
colorlog>=6.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Type checking and linting
mypy>=1.5.0
black>=23.7.0
flake8>=6.1.0

# Optional: vLLM for local embedding models
# vllm>=0.8.3
# transformers>=4.56.1
# torch>=2.0.0
